display temperature and change types to uint

+++ Inc/circularbuf.h
@@ -9,7 +9,7 @@
 
 typedef struct
 {
-	int8_t *buff_array;
+	uint8_t *buff_array;
 	uint8_t head ;
 	uint8_t tail;
 	uint8_t size;
@@ -19,8 +19,8 @@ circular_buffer_t* circular_buf_initialize(uint8_t);
 
 int circular_buf_empty(circular_buffer_t*);
 
-int circular_buf_get(circular_buffer_t*);
+uint8_t circular_buf_get(circular_buffer_t*);
 
-void circular_buf_put(circular_buffer_t*, int8_t);
+void circular_buf_put(circular_buffer_t*, uint8_t);
 
 #endif
