+++ Src/circularbuf.c
@@ -3,14 +3,14 @@
 circular_buffer_t* circular_buf_initialize(uint8_t buff_size)
 {
 	circular_buffer_t *buff = (circular_buffer_t *) malloc(sizeof(circular_buffer_t));
-	buff->buff_array = (int8_t *) malloc(sizeof(int8_t)*buff_size);
+	buff->buff_array = (uint8_t *) malloc(sizeof(uint8_t)*buff_size);
 	buff->head = 0;
 	buff->tail = 0;
 	buff->size = buff_size;
 	return buff;
 }
 
-void circular_buf_put(circular_buffer_t* buff, int8_t value)
+void circular_buf_put(circular_buffer_t* buff, uint8_t value)
 {
 	buff->buff_array[buff->head] = value;
 	if(buff->head < buff->size) 
@@ -23,9 +23,9 @@ void circular_buf_put(circular_buffer_t* buff, int8_t value)
 	}
 }
 
-int circular_buf_get(circular_buffer_t* buff)
+uint8_t circular_buf_get(circular_buffer_t* buff)
 {
-	int value = buff->buff_array[buff->tail];
+	uint8_t value = buff->buff_array[buff->tail];
 	if(buff->tail < buff->size) 
 	{
 		buff->tail++;
